Fixed libcurl dependency problem when compiling the program
Added the ability to load primer3settings.conf through environmental variables
Ouput apache logs to stdout and stderr
Changed the entrypoint to tail stdout instead of null
Removed sudo from Dockerfile since we are running it as root anyway
Added sample files. These seems to be in the image from docker hub but not when you build the image from scratch

Changed the docker-compose file to build from the dockerfile instead of pulling a premade image. I commented out the
original line to pull from docker hub so people can switch to that if they want to. It might be a good idea to build from the dockerfile
since it's available in the repo and so that individuals can easily debug their builds. The downside is the time to build.